requirements.txt
cachelib==0.1.1
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
Flask==1.1.2
Flask-Cors==3.0.10
Flask-Session==0.3.2
gunicorn==20.0.4
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
# numpy==1.19.5
pylast==4.1.0
python-dotenv==0.15.0
requests==2.25.1
six==1.15.0
spotipy==2.16.1
urllib3==1.26.3
uuid==1.30
Werkzeug==1.0.1
